Great Stale Needle Cleanup 2018: anaconda
It's my second least favorite day of the year again: Stale
Needle Cleanup Day!

This should get rid of all anaconda needles that definitely are
not being used any more. A few borderline cases (where I'm not
100% sure if they may still be useful in odd corner cases, like
running universal tests on non-server images, and runs outside
of the US, and stuff) are kept around.

Cleanup of all non-anaconda needles will be in the next commit.

Signed-off-by: Adam Williamson <awilliam@redhat.com>