Don't install updates on support server if versions don't match
I caught a few failed support_server update tests which failed
because packages from the update weren't installed; they weren't
installed because the update was for F28, but the support server
runs on an F29 disk image. Obviously we should skip the 'install
packages from the update' step in the support_server test unless
the versions match.

Arguably we could *never* install updates in support_server, but
I think it's actually valuable to know if the update breaks
support_server if we *are* on the right version.

Signed-off-by: Adam Williamson <awilliam@redhat.com>