Re-enable memory_check test (on Everything, with WS packages)
We were only running this test on the Workstation boot ISO...
then that was removed, and so we actually never ran this test at
all any more. We should run it on at least one image as check-
compose uses the output to check whether installer memory usage
has increased. I initially ran it on Workstation as that installs
the most packages; so let's run it on Everything but use the
Workstation package set.

Signed-off-by: Adam Williamson <awilliam@redhat.com>