Fix firefox addon_success needle
I updated this last month without looking closely enough; one of
the matches was bogus. We now don't seem to get the 'added to
Firefox' popup any more, so we can't match on it, but I guess we
can match on the 'remove' button as an indicator that the add
worked.

Signed-off-by: Adam Williamson <awilliam@redhat.com>