Update F32 background needles
This adds a needle for the 'day' background from 32.1.1, which
is the non-animated default, and renames the needle for the
background from 32.0.0 to 'old' as it's no longer present at all
in the new build. Depending on whether animation actually works
after the update we may later need to add needles for the other
images plus possibly some transition needles too. Sigh.

Signed-off-by: Adam Williamson <awilliam@redhat.com>