Add a helper script for moving needle files
I actually wrote this a while back but have just been keeping it
on my local machine till now; figured I'd clean it up a bit and
add it to the repo. It's just a little helper for moving needles
into subdirectories, it's useful if you do a big needle retake
and have dozens of needles you need to put in the right subdirs,
saves figuring them out for yourself.

Signed-off-by: Adam Williamson <awilliam@redhat.com>