Tweak NFS repo setup in _support_server to copy all files
`cp -R foo/*` doesn't get all files in `foo/`, it misses hidden
files. This turns out to be a problem with recent anaconda, as
it expects to find a .treeinfo file here. So, let's use rsync.
We could probably do this with cp too but I can't think of the
right arguments right now...

Signed-off-by: Adam Williamson <awilliam@redhat.com>